+++ core/control/motion_primitives/motion_primitives_common.hpp
@@ -33,7 +33,7 @@ namespace motion_primitives_common
 
   bool has_elapsed(std::uint32_t now_ms, std::uint32_t target_ms);
   std::int64_t abs_i64(std::int64_t value);
-  void send_motion_command(bool drive_enabled, std::int16_t linear_velocity_mm_s, std::int16_t yaw_rate_mdeg_s, std::uint32_t now_ms);
+  void send_motion_command(bool drive_enabled, std::int32_t linear_velocity_mm_s, std::int32_t yaw_rate_mdeg_s, std::uint32_t now_ms);
   void stop_motion(std::uint32_t now_ms);
   std::int64_t squared_distance_um(const local_positioning::snapshot &start_pose, const local_positioning::snapshot &current_pose);
   bool has_reached_forward_target(const local_positioning::snapshot &start_pose, const local_positioning::snapshot &current_pose, std::int64_t target_distance_um);
